// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const FileType = {
  PDF: 'PDF',
  DOCX: 'DOCX',
  MARKDOWN: 'MARKDOWN',
  TXT: 'TXT',
  CSV: 'CSV'
} as const

export type FileType = (typeof FileType)[keyof typeof FileType]


export const DocumentStatus = {
  PENDING: 'PENDING',
  READY: 'READY',
  FAILED: 'FAILED'
} as const

export type DocumentStatus = (typeof DocumentStatus)[keyof typeof DocumentStatus]


export const Role = {
  USER: 'USER',
  ASSISTANT: 'ASSISTANT'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Plan = {
  FREE: 'FREE',
  PRO: 'PRO',
  TEAM: 'TEAM'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const SubStatus = {
  ACTIVE: 'ACTIVE',
  CANCELED: 'CANCELED',
  PAST_DUE: 'PAST_DUE',
  TRIALING: 'TRIALING'
} as const

export type SubStatus = (typeof SubStatus)[keyof typeof SubStatus]